French translation
Hello!

This is my first contribution on GitHub.
I have discovered CommaFeed last week and I love it! Thank you for this. After exploring pages of the site, I saw I could contribute for translations… then I decided to contributed.
My native language is French. My English is not perfect but my experience of the Internet as a user allows me to translate easily websites from English to French.
I am pretty sure of the accuracy of all the edited translations, except for one: "filtering_expression_help".
I wrote a comment after each modification and each time I thought it was useful  (I am not a developer but I think it is correct to use // in the file).

I hope this contribution will help you. And thank you again for CommaFeed.

Adextre